VisionX/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==10.0.0
Flask-SQLAlchemy==3.0.5
xx